Delete unnecessary imports, fix product general form and adjust subcategories to category-control and delete category dialog

+++ src/app/components/features/admin-panel/category-control/category-control.ts
@@ -1,15 +1,18 @@
-import {Component, computed, inject, signal} from '@angular/core';
+import {Component, inject, OnInit, signal} from '@angular/core';
 import {Button} from '../../../shared/button/button';
 import {TranslatePipe} from '@ngx-translate/core';
 import {CategoryService} from '../../../../core/services/category-service/category.service';
 import {DomSanitizer} from '@angular/platform-browser';
-import {RouterLink} from '@angular/router';
-import {MainCategoryDialog} from './main-category-dialog/main-category-dialog';
+import {ActivatedRoute, RouterLink} from '@angular/router';
+import {CategoryDialog} from './main-category-dialog/category-dialog';
 import {CategoryDTO} from '../../../../core/models/category/CategoryDTO';
-import {ConfirmDialog} from '../../../shared/confirm-dialog/confirm-dialog';
 import {LoadingDialog} from '../../../shared/loading-dialog/loading-dialog';
 import {AdminService} from '../../../../core/services/admin-service/admin.service';
 import {finalize} from 'rxjs';
+import {CategoryView} from '../../../../core/models/category/CategoryView';
+import {Dialog} from '../../../shared/dialog/dialog';
+import {FormBuilder, FormControl, ReactiveFormsModule} from '@angular/forms';
+import {DeleteAdminPanelCategoryDTO} from '../../../../core/models/category/DeleteAdminPanelCategoryDTO';
 
 @Component({
   selector: 'app-category-control',
@@ -17,24 +20,22 @@ import {finalize} from 'rxjs';
     Button,
     TranslatePipe,
     RouterLink,
-    MainCategoryDialog,
-    ConfirmDialog,
+    CategoryDialog,
     LoadingDialog,
+    Dialog,
+    ReactiveFormsModule,
   ],
   templateUrl: './category-control.html',
   styleUrl: './category-control.css',
 })
-export class CategoryControl {
+export class CategoryControl implements OnInit {
   private _categoryService = inject(CategoryService);
   private _adminService = inject(AdminService);
   private _sanitizer = inject(DomSanitizer);
+  private _route = inject(ActivatedRoute);
+  private _formBuilder = inject(FormBuilder);
 
-  readonly categories = computed(() =>
-    this._categoryService.mainCategories().map(category => ({
-      ...category,
-      safeSvg: category.svgImage ? this._sanitizer.bypassSecurityTrustHtml(category.svgImage) : ''
-    }))
-  );
+  readonly categories = signal<CategoryView[]>([]);
 
   readonly isOpenDialog = signal<boolean>(false);
   readonly isOpenDeleteDialog = signal<boolean>(false);
@@ -42,6 +43,19 @@ export class CategoryControl {
   readonly loading = signal<boolean>(false);
 
   readonly deleteCategoryId = signal<number | null>(null);
+  readonly parentCategoryId = signal<number | null>(null);
+
+  readonly deleteForm = this._formBuilder.group({
+    deleteFullTree: new FormControl<boolean>(false),
+  });
+
+  ngOnInit(): void {
+    this._route.queryParams.subscribe(params => {
+      const id = params['id'] ? +params['id'] : null;
+      this.parentCategoryId.set(id);
+      this.loadCategories(id);
+    });
+  }
 
   openDialog(category?: CategoryDTO): void {
     if(category) {
@@ -59,19 +73,34 @@ export class CategoryControl {
   }
 
   closeDeleteDialog(): void {
+    this.deleteForm.reset({
+      deleteFullTree: false,
+    });
     this.deleteCategoryId.set(null);
     this.isOpenDeleteDialog.set(false);
   }
 
   confirmDeleteCategory(): void {
+    if (this.deleteForm.invalid) {
+      this.deleteForm.markAllAsTouched();
+      return;
+    }
+
     const categoryId = this.deleteCategoryId();
     if(!categoryId) return;
 
+    const payload : DeleteAdminPanelCategoryDTO = {
+      deleteFullTree: this.deleteForm.controls.deleteFullTree.value ?? false
+    };
+
     this.loading.set(true);
 
-    this._adminService.deleteCategory(categoryId).pipe(finalize(() => {
+    this._adminService.deleteCategory(categoryId, payload).pipe(finalize(() => {
       this.loading.set(false);
     })).subscribe(() => {
+      this.deleteForm.reset({
+        deleteFullTree: false,
+      });
       this.refreshCategories();
     });
 
@@ -79,8 +108,31 @@ export class CategoryControl {
   }
 
   refreshCategories(): void {
-    this._categoryService.loadMainCategories().subscribe(() => {
-      console.log(this._categoryService.mainCategories());
-    });
+    this.loadCategories(this.parentCategoryId());
+  }
+
+  private loadCategories(parentId: number | null) {
+    this.loading.set(true);
+
+    if(!parentId){
+      this._categoryService.loadMainCategories().pipe(
+        finalize(() => this.loading.set(false))
+      ).subscribe(categories => {
+        this.categories.set(categories.map(category => ({
+          ...category,
+          safeSvg: category.svgImage ? this._sanitizer.bypassSecurityTrustHtml(category.svgImage) : null
+        })));
+      });
+    }
+    else {
+      this._categoryService.loadSubcategories(parentId).pipe(
+        finalize(() => this.loading.set(false))
+      ).subscribe(categories => {
+        this.categories.set(categories.map(category => ({
+          ...category,
+          safeSvg: category.svgImage ? this._sanitizer.bypassSecurityTrustHtml(category.svgImage) : null
+        })));
+      });
+    }
   }
 }
